Initial commit: Basic report functionality.
This should be approximately everything for generating a report (and printing the String). However the IDE deleted a couple of files that I had to manually restore so it's possible there are discrepancies that I'll need to fix.

This also includes logic to determine whether the previous month's budget is ready to be generated. Up next is functionality to email the report.